Python/Django/Supera-Inovacao-Tecnologia-exercise/backend/requirements.txt
python-dotenv
django
djangorestframework
django-cors-headers
psycopg2
django-rest-knox